+++ src/components/TilemapUtil.ts
@@ -28,7 +28,6 @@ export default class TilemapUtil {
             const _obj = obj as any;
             return _obj.gid === GID;
 		});
-			// TODO: I shouldn't have to name these anymore, since they have unique GIDs.
 
         return new Phaser.Math.Vector2(point.x!, point.y!)
     }
@@ -45,6 +44,21 @@ export default class TilemapUtil {
 		{
             const _obj = obj as any;
 			if(obj.name === objectName)
+            {
+                flip = _obj.flippedHorizontal;
+            }
+		});
+
+        return flip;
+    }
+
+    public static getObjectFlipByGID(GID: number, tilemap:Phaser.Tilemaps.Tilemap): boolean
+    {
+        let flip = false;
+        const point = tilemap.findObject('elements', function (obj)
+		{
+            const _obj = obj as any;
+            if (_obj.gid === GID)
             {
                 flip = _obj.flippedHorizontal;
             }