+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Enemy settings in the tilemap editor have been unified as custom properties.
 - VisionPoly no longer needs to be named such in the tilemap, as it's identified by being a polygon object.
 - Starting point no longer needs to be named such, as it's identified by GID.
+- Updated all level tilemaps to work with updated loading methods.
 
 ### Added